requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
django==5.0.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.12.2
    # via asgiref
tzdata==2024.1
    # via django